requirements.txt
algoliasearch          3.0.0
blinker                1.9.0
certifi                2025.1.31
cffi                   2.0.0
charset-normalizer     3.4.5
click                  8.3.1
colorama               0.4.6
colorlog               6.9.0
cryptography           44.0.3
decorator              5.2.1
Flask                  3.1.1
greenlet               3.3.2
gunicorn               23.0.0
idna                   3.11
iniconfig              2.3.0
itsdangerous           2.2.0
Jinja2                 3.1.6
MarkupSafe             3.0.3
mysql-connector-python 9.3.0
packaging              26.0
pip                    23.2.1
pluggy                 1.6.0
py                     1.11.0
pycparser              3.0
PyMySQL                1.1.1
pytest                 8.3.5
pytest-flask           1.3.0
python-dateutil        2.9.0
python-dotenv          1.1.1
pytz                   2025.1
redis                  7.3.0
requests               2.32.4
retry                  0.9.2
setuptools             65.5.0
six                    1.17.0
SQLAlchemy             2.0.41
tenacity               9.1.2
typing_extensions      4.15.0
urllib3                2.4.0
Werkzeug               3.1.3